Fix SettingsRoute and DevicesRoute
